refactor(privacy_preserving_circuit): extract output module
Refs: #454
